Migrate Custom threshold duplicated tests to the deployment agnostic framework (#198691)
Closes #183378
Closes #179095

## Summary
This PR moves the Custom threshold rule duplicated API integration tests
to the deployment agnostic test.

## How to run

To run serverless
```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/serverless/oblt.serverless.config.ts --grep="Custom Threshold rule"
```

To run stateful
```
node scripts/functional_tests_server --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts
node scripts/functional_test_runner --config x-pack/test/api_integration/deployment_agnostic/configs/stateful/oblt.stateful.config.ts --grep="Custom Threshold rule"
```

### TODO

- [x] Test in MKI before merging

#### How to run tests on MKI

According to this
[discussion](https://github.com/elastic/observability-dev/issues/3519#issuecomment-2379914274),
we should test in MKI environment before merging. For details on how to
run in MKI, see [this section of the
document](https://docs.google.com/document/d/1tiax7xoDYwFXYZjRTgVKkVMjN-SQzBWk4yn1JY6Z5UY/edit#heading=h.ece2z8p74izh)
and [this
readme](https://github.com/elastic/kibana/blob/main/x-pack/test_serverless/README.md#run-tests-on-mki).

---------

(cherry picked from commit b585ca658aafeac2d83b5cd2b3a3c2c917c90d01)

# Conflicts:
#	x-pack/test_serverless/api_integration/test_suites/observability/index.feature_flags.ts
#	x-pack/test_serverless/tsconfig.json
