--- x-pack/test/alerting_api_integration/observability/custom_threshold_rule/rate_bytes_fired.ts
@@ -1,283 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { cleanup, generate, Dataset, PartialConfig } from '@kbn/data-forge';
-import { Aggregators } from '@kbn/observability-plugin/common/custom_threshold_rule/types';
-import { COMPARATORS } from '@kbn/alerting-comparators';
-import { FIRED_ACTIONS_ID } from '@kbn/observability-plugin/server/lib/rules/custom_threshold/constants';
-import expect from '@kbn/expect';
-import { OBSERVABILITY_THRESHOLD_RULE_TYPE_ID } from '@kbn/rule-data-utils';
-import { createIndexConnector, createRule } from '../helpers/alerting_api_helper';
-import { createDataView, deleteDataView } from '../helpers/data_view';
-import {
-  waitForAlertInIndex,
-  waitForDocumentInIndex,
-  waitForRuleStatus,
-} from '../helpers/alerting_wait_for_helpers';
-import { FtrProviderContext } from '../../common/ftr_provider_context';
-import { ActionDocument } from './typings';
-
-// eslint-disable-next-line import/no-default-export
-export default function ({ getService }: FtrProviderContext) {
-  const esClient = getService('es');
-  const supertest = getService('supertest');
-  const esDeleteAllIndices = getService('esDeleteAllIndices');
-  const logger = getService('log');
-  const retryService = getService('retry');
-
-  describe('Custom Threshold rule RATE - GROUP_BY - BYTES - FIRED', () => {
-    const CUSTOM_THRESHOLD_RULE_ALERT_INDEX = '.alerts-observability.threshold.alerts-default';
-    const ALERT_ACTION_INDEX = 'alert-action-threshold';
-    const DATE_VIEW = 'kbn-data-forge-fake_hosts.fake_hosts-*';
-    const DATA_VIEW_ID = 'data-view-id';
-    let dataForgeConfig: PartialConfig;
-    let dataForgeIndices: string[];
-    let actionId: string;
-    let ruleId: string;
-    let alertId: string;
-
-    before(async () => {
-      dataForgeConfig = {
-        schedule: [
-          {
-            template: 'good',
-            start: 'now-10m',
-            end: 'now+5m',
-            metrics: [
-              { name: 'system.network.in.bytes', method: 'linear', start: 0, end: 54000000 },
-            ],
-          },
-        ],
-        indexing: {
-          dataset: 'fake_hosts' as Dataset,
-          eventsPerCycle: 1,
-          interval: 10000,
-          alignEventsToInterval: true,
-        },
-      };
-      dataForgeIndices = await generate({ client: esClient, config: dataForgeConfig, logger });
-      await waitForDocumentInIndex({
-        esClient,
-        indexName: dataForgeIndices.join(','),
-        docCountTarget: 270,
-        retryService,
-        logger,
-      });
-      await createDataView({
-        supertest,
-        name: DATE_VIEW,
-        id: DATA_VIEW_ID,
-        title: DATE_VIEW,
-        logger,
-      });
-    });
-
-    after(async () => {
-      await supertest.delete(`/api/alerting/rule/${ruleId}`).set('kbn-xsrf', 'foo');
-      await supertest.delete(`/api/actions/connector/${actionId}`).set('kbn-xsrf', 'foo');
-      await esClient.deleteByQuery({
-        index: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
-        query: { term: { 'kibana.alert.rule.uuid': ruleId } },
-      });
-      await esClient.deleteByQuery({
-        index: '.kibana-event-log-*',
-        query: { term: { 'kibana.alert.rule.consumer': 'logs' } },
-      });
-      await deleteDataView({
-        supertest,
-        id: DATA_VIEW_ID,
-        logger,
-      });
-      await esDeleteAllIndices([ALERT_ACTION_INDEX, ...dataForgeIndices]);
-      await cleanup({ client: esClient, config: dataForgeConfig, logger });
-    });
-
-    describe('Rule creation', () => {
-      it('creates rule successfully', async () => {
-        actionId = await createIndexConnector({
-          supertest,
-          name: 'Index Connector: Threshold API test',
-          indexName: ALERT_ACTION_INDEX,
-          logger,
-        });
-
-        const createdRule = await createRule({
-          supertest,
-          logger,
-          esClient,
-          tags: ['observability'],
-          consumer: 'logs',
-          name: 'Threshold rule',
-          ruleTypeId: OBSERVABILITY_THRESHOLD_RULE_TYPE_ID,
-          params: {
-            criteria: [
-              {
-                comparator: COMPARATORS.GREATER_THAN_OR_EQUALS,
-                threshold: [50_000],
-                timeSize: 1,
-                timeUnit: 'm',
-                metrics: [
-                  { name: 'A', field: 'system.network.in.bytes', aggType: Aggregators.RATE },
-                ],
-              },
-            ],
-            alertOnNoData: true,
-            alertOnGroupDisappear: true,
-            searchConfiguration: {
-              query: {
-                query: '',
-                language: 'kuery',
-              },
-              index: DATA_VIEW_ID,
-            },
-            groupBy: ['host.name', 'container.id'],
-          },
-          actions: [
-            {
-              group: FIRED_ACTIONS_ID,
-              id: actionId,
-              params: {
-                documents: [
-                  {
-                    ruleType: '{{rule.type}}',
-                    alertDetailsUrl: '{{context.alertDetailsUrl}}',
-                    reason: '{{context.reason}}',
-                    value: '{{context.value}}',
-                    host: '{{context.host}}',
-                    group: '{{context.group}}',
-                  },
-                ],
-              },
-              frequency: {
-                notify_when: 'onActionGroupChange',
-                throttle: null,
-                summary: false,
-              },
-            },
-          ],
-        });
-        ruleId = createdRule.id;
-        expect(ruleId).not.to.be(undefined);
-      });
-
-      it('should be active', async () => {
-        const executionStatus = await waitForRuleStatus({
-          id: ruleId,
-          expectedStatus: 'active',
-          supertest,
-          retryService,
-          logger,
-        });
-        expect(executionStatus.status).to.be('active');
-      });
-
-      it('should set correct information in the alert document', async () => {
-        const resp = await waitForAlertInIndex({
-          esClient,
-          indexName: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
-          ruleId,
-          retryService,
-          logger,
-        });
-        alertId = (resp.hits.hits[0]._source as any)['kibana.alert.uuid'];
-
-        expect(resp.hits.hits[0]._source).property(
-          'kibana.alert.rule.category',
-          'Custom threshold'
-        );
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', 'logs');
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.name', 'Threshold rule');
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.producer', 'observability');
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.revision', 0);
-        expect(resp.hits.hits[0]._source).property(
-          'kibana.alert.rule.rule_type_id',
-          'observability.rules.custom_threshold'
-        );
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.uuid', ruleId);
-        expect(resp.hits.hits[0]._source).property('kibana.space_ids').contain('default');
-        expect(resp.hits.hits[0]._source)
-          .property('kibana.alert.rule.tags')
-          .contain('observability');
-        expect(resp.hits.hits[0]._source).property(
-          'kibana.alert.action_group',
-          'custom_threshold.fired'
-        );
-        expect(resp.hits.hits[0]._source).property('tags').contain('observability');
-        expect(resp.hits.hits[0]._source).property(
-          'kibana.alert.instance.id',
-          'host-0,container-0'
-        );
-        expect(resp.hits.hits[0]._source).property('kibana.alert.workflow_status', 'open');
-        expect(resp.hits.hits[0]._source).property('event.kind', 'signal');
-        expect(resp.hits.hits[0]._source).property('event.action', 'open');
-
-        expect(resp.hits.hits[0]._source).property('host.name', 'host-0');
-        expect(resp.hits.hits[0]._source)
-          .property('host.mac')
-          .eql(['00-00-5E-00-53-23', '00-00-5E-00-53-24']);
-        expect(resp.hits.hits[0]._source).property('container.id', 'container-0');
-        expect(resp.hits.hits[0]._source).property('container.name', 'container-name');
-        expect(resp.hits.hits[0]._source).not.property('container.cpu');
-
-        expect(resp.hits.hits[0]._source)
-          .property('kibana.alert.group')
-          .eql([
-            {
-              field: 'host.name',
-              value: 'host-0',
-            },
-            {
-              field: 'container.id',
-              value: 'container-0',
-            },
-          ]);
-
-        expect(resp.hits.hits[0]._source)
-          .property('kibana.alert.rule.parameters')
-          .eql({
-            criteria: [
-              {
-                comparator: '>=',
-                threshold: [50_000],
-                timeSize: 1,
-                timeUnit: 'm',
-                metrics: [{ name: 'A', field: 'system.network.in.bytes', aggType: 'rate' }],
-              },
-            ],
-            alertOnNoData: true,
-            alertOnGroupDisappear: true,
-            searchConfiguration: { index: 'data-view-id', query: { query: '', language: 'kuery' } },
-            groupBy: ['host.name', 'container.id'],
-          });
-      });
-
-      it('should set correct action variables', async () => {
-        const resp = await waitForDocumentInIndex<ActionDocument>({
-          esClient,
-          indexName: ALERT_ACTION_INDEX,
-          retryService,
-          logger,
-        });
-
-        expect(resp.hits.hits[0]._source?.ruleType).eql('observability.rules.custom_threshold');
-        expect(resp.hits.hits[0]._source?.alertDetailsUrl).eql(
-          `https://localhost:5601/app/observability/alerts/${alertId}`
-        );
-        expect(resp.hits.hits[0]._source?.reason).eql(
-          `Rate of system.network.in.bytes is 60 kB/s, above or equal the threshold of 50 kB/s. (duration: 1 min, data view: kbn-data-forge-fake_hosts.fake_hosts-*, group: host-0,container-0)`
-        );
-        expect(resp.hits.hits[0]._source?.value).eql('60 kB/s');
-        expect(resp.hits.hits[0]._source?.host).eql(
-          '{"name":"host-0","mac":["00-00-5E-00-53-23","00-00-5E-00-53-24"]}'
-        );
-        expect(resp.hits.hits[0]._source?.group).eql(
-          '{"field":"host.name","value":"host-0"},{"field":"container.id","value":"container-0"}'
-        );
-      });
-    });
-  });
-}
