+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/alerting/custom_threshold/avg_pct_no_data.ts
@@ -7,133 +7,109 @@
 
 import { omit } from 'lodash';
 import expect from '@kbn/expect';
-import { cleanup, generate, Dataset, PartialConfig } from '@kbn/data-forge';
 import { Aggregators } from '@kbn/observability-plugin/common/custom_threshold_rule/types';
-import { FIRED_ACTIONS_ID } from '@kbn/observability-plugin/server/lib/rules/custom_threshold/constants';
+import { NO_DATA_ACTIONS_ID } from '@kbn/observability-plugin/server/lib/rules/custom_threshold/constants';
 import { OBSERVABILITY_THRESHOLD_RULE_TYPE_ID } from '@kbn/rule-data-utils';
 import { parseSearchParams } from '@kbn/share-plugin/common/url_service';
 import { COMPARATORS } from '@kbn/alerting-comparators';
-import { createIndexConnector, createRule } from '../helpers/alerting_api_helper';
-import { createDataView, deleteDataView } from '../helpers/data_view';
-import {
-  waitForAlertInIndex,
-  waitForDocumentInIndex,
-  waitForRuleStatus,
-} from '../helpers/alerting_wait_for_helpers';
-import { FtrProviderContext } from '../../common/ftr_provider_context';
+import { kbnTestConfig } from '@kbn/test';
+import type { InternalRequestHeader, RoleCredentials } from '@kbn/ftr-common-functional-services';
+import { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
 import { ISO_DATE_REGEX } from './constants';
-import { ActionDocument, LogsExplorerLocatorParsedParams } from './typings';
+import { ActionDocument, LogsExplorerLocatorParsedParams } from './types';
 
-// eslint-disable-next-line import/no-default-export
-export default function ({ getService }: FtrProviderContext) {
+export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
   const esClient = getService('es');
-  const supertest = getService('supertest');
+  const supertestWithoutAuth = getService('supertestWithoutAuth');
+  const alertingApi = getService('alertingApi');
+  const dataViewApi = getService('dataViewApi');
   const esDeleteAllIndices = getService('esDeleteAllIndices');
-  const logger = getService('log');
-  const retryService = getService('retry');
+  const samlAuth = getService('samlAuth');
+  let roleAuthc: RoleCredentials;
+  let internalReqHeader: InternalRequestHeader;
+  const config = getService('config');
+  const isServerless = config.get('serverless');
+  const expectedConsumer = isServerless ? 'observability' : 'logs';
 
-  describe('Custom Threshold rule - DOCUMENTS_COUNT - FIRED', () => {
+  describe('AVG - PCT - NoData', () => {
     const CUSTOM_THRESHOLD_RULE_ALERT_INDEX = '.alerts-observability.threshold.alerts-default';
     const ALERT_ACTION_INDEX = 'alert-action-threshold';
-    const DATA_VIEW = 'kbn-data-forge-fake_hosts.fake_hosts-*';
-    const DATA_VIEW_ID = 'data-view-id';
-    const DATA_VIEW_NAME = 'data-view-name';
-    let dataForgeConfig: PartialConfig;
-    let dataForgeIndices: string[];
+    const DATA_VIEW_NAME = 'no-data-pattern-name';
+    const DATA_VIEW_ID = 'data-view-id-no-data';
+    const DATA_VIEW_TITLE = 'no-data-pattern-title';
     let actionId: string;
     let ruleId: string;
     let alertId: string;
 
     before(async () => {
-      dataForgeConfig = {
-        schedule: [
-          {
-            template: 'good',
-            start: 'now-10m',
-            end: 'now+5m',
-            metrics: [
-              { name: 'system.cpu.user.pct', method: 'linear', start: 2.5, end: 2.5 },
-              { name: 'system.cpu.total.pct', method: 'linear', start: 0.5, end: 0.5 },
-              { name: 'system.cpu.total.norm.pct', method: 'linear', start: 0.8, end: 0.8 },
-            ],
-          },
-        ],
-        indexing: {
-          dataset: 'fake_hosts' as Dataset,
-          eventsPerCycle: 1,
-          interval: 60000,
-          alignEventsToInterval: true,
-        },
-      };
-      dataForgeIndices = await generate({ client: esClient, config: dataForgeConfig, logger });
-      await waitForDocumentInIndex({
-        esClient,
-        indexName: dataForgeIndices.join(','),
-        docCountTarget: 45,
-        retryService,
-        logger,
-      });
-      await createDataView({
-        supertest,
+      roleAuthc = await samlAuth.createM2mApiKeyWithRoleScope('admin');
+      internalReqHeader = samlAuth.getInternalRequestHeader();
+      await dataViewApi.create({
         name: DATA_VIEW_NAME,
         id: DATA_VIEW_ID,
-        title: DATA_VIEW,
-        logger,
+        title: DATA_VIEW_TITLE,
+        roleAuthc,
       });
     });
 
     after(async () => {
-      await supertest.delete(`/api/alerting/rule/${ruleId}`).set('kbn-xsrf', 'foo');
-      await supertest.delete(`/api/actions/connector/${actionId}`).set('kbn-xsrf', 'foo');
+      await supertestWithoutAuth
+        .delete(`/api/alerting/rule/${ruleId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(internalReqHeader);
+      await supertestWithoutAuth
+        .delete(`/api/actions/connector/${actionId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(internalReqHeader);
       await esClient.deleteByQuery({
         index: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
         query: { term: { 'kibana.alert.rule.uuid': ruleId } },
+        conflicts: 'proceed',
       });
       await esClient.deleteByQuery({
         index: '.kibana-event-log-*',
-        query: { term: { 'kibana.alert.rule.consumer': 'logs' } },
+        query: { term: { 'rule.id': ruleId } },
+        conflicts: 'proceed',
       });
-      await deleteDataView({
-        supertest,
+      await dataViewApi.delete({
         id: DATA_VIEW_ID,
-        logger,
+        roleAuthc,
       });
-      await esDeleteAllIndices([ALERT_ACTION_INDEX, ...dataForgeIndices]);
-      await cleanup({ client: esClient, config: dataForgeConfig, logger });
+      await esDeleteAllIndices([ALERT_ACTION_INDEX]);
+      await samlAuth.invalidateM2mApiKeyWithRoleScope(roleAuthc);
     });
 
     describe('Rule creation', () => {
       it('creates rule successfully', async () => {
-        actionId = await createIndexConnector({
-          supertest,
+        actionId = await alertingApi.createIndexConnector({
+          roleAuthc,
           name: 'Index Connector: Threshold API test',
           indexName: ALERT_ACTION_INDEX,
-          logger,
         });
 
-        const createdRule = await createRule({
-          supertest,
-          logger,
-          esClient,
+        const createdRule = await alertingApi.createRule({
+          roleAuthc,
           tags: ['observability'],
-          consumer: 'logs',
+          consumer: expectedConsumer,
           name: 'Threshold rule',
           ruleTypeId: OBSERVABILITY_THRESHOLD_RULE_TYPE_ID,
           params: {
             criteria: [
               {
-                comparator: COMPARATORS.NOT_BETWEEN,
-                threshold: [1, 2],
-                timeSize: 1,
+                comparator: COMPARATORS.GREATER_THAN,
+                threshold: [0.5],
+                timeSize: 5,
                 timeUnit: 'm',
-                metrics: [{ name: 'A', filter: 'container.id:*', aggType: Aggregators.COUNT }],
+                metrics: [
+                  { name: 'A', field: 'system.cpu.user.pct', aggType: Aggregators.AVERAGE },
+                ],
               },
             ],
             alertOnNoData: true,
             alertOnGroupDisappear: true,
             searchConfiguration: {
               query: {
-                query: 'host.name:*',
+                query: '',
                 language: 'kuery',
               },
               index: DATA_VIEW_ID,
@@ -141,7 +117,7 @@ export default function ({ getService }: FtrProviderContext) {
           },
           actions: [
             {
-              group: FIRED_ACTIONS_ID,
+              group: NO_DATA_ACTIONS_ID,
               id: actionId,
               params: {
                 documents: [
@@ -167,23 +143,24 @@ export default function ({ getService }: FtrProviderContext) {
       });
 
       it('should be active', async () => {
-        const executionStatus = await waitForRuleStatus({
-          id: ruleId,
+        const executionStatus = await alertingApi.waitForRuleStatus({
+          roleAuthc,
+          ruleId,
           expectedStatus: 'active',
-          supertest,
-          retryService,
-          logger,
         });
-        expect(executionStatus.status).to.be('active');
+        expect(executionStatus).to.be('active');
+      });
+
+      it('should find the created rule with correct information about the consumer', async () => {
+        const match = await alertingApi.findInRules(roleAuthc, ruleId);
+        expect(match).not.to.be(undefined);
+        expect(match.consumer).to.be(expectedConsumer);
       });
 
       it('should set correct information in the alert document', async () => {
-        const resp = await waitForAlertInIndex({
-          esClient,
+        const resp = await alertingApi.waitForAlertInIndex({
           indexName: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
           ruleId,
-          retryService,
-          logger,
         });
         alertId = (resp.hits.hits[0]._source as any)['kibana.alert.uuid'];
 
@@ -191,7 +168,7 @@ export default function ({ getService }: FtrProviderContext) {
           'kibana.alert.rule.category',
           'Custom threshold'
         );
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', 'logs');
+        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', expectedConsumer);
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.name', 'Threshold rule');
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.producer', 'observability');
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.revision', 0);
@@ -206,7 +183,7 @@ export default function ({ getService }: FtrProviderContext) {
           .contain('observability');
         expect(resp.hits.hits[0]._source).property(
           'kibana.alert.action_group',
-          'custom_threshold.fired'
+          'custom_threshold.nodata'
         );
         expect(resp.hits.hits[0]._source).property('tags').contain('observability');
         expect(resp.hits.hits[0]._source).property('kibana.alert.instance.id', '*');
@@ -214,46 +191,42 @@ export default function ({ getService }: FtrProviderContext) {
         expect(resp.hits.hits[0]._source).property('event.kind', 'signal');
         expect(resp.hits.hits[0]._source).property('event.action', 'open');
 
-        expect(resp.hits.hits[0]._source).not.have.property('kibana.alert.group');
-        expect(resp.hits.hits[0]._source).not.have.property('kibana.alert.evaluation.threshold');
         expect(resp.hits.hits[0]._source)
           .property('kibana.alert.rule.parameters')
           .eql({
             criteria: [
               {
-                comparator: COMPARATORS.NOT_BETWEEN,
-                threshold: [1, 2],
-                timeSize: 1,
+                comparator: '>',
+                threshold: [0.5],
+                timeSize: 5,
                 timeUnit: 'm',
-                metrics: [{ name: 'A', filter: 'container.id:*', aggType: 'count' }],
+                metrics: [{ name: 'A', field: 'system.cpu.user.pct', aggType: 'avg' }],
               },
             ],
             alertOnNoData: true,
             alertOnGroupDisappear: true,
             searchConfiguration: {
-              index: 'data-view-id',
-              query: { query: 'host.name:*', language: 'kuery' },
+              index: 'data-view-id-no-data',
+              query: { query: '', language: 'kuery' },
             },
           });
       });
 
       it('should set correct action variables', async () => {
-        const resp = await waitForDocumentInIndex<ActionDocument>({
-          esClient,
+        const resp = await alertingApi.waitForDocumentInIndex<ActionDocument>({
           indexName: ALERT_ACTION_INDEX,
-          retryService,
-          logger,
+          docCountTarget: 1,
         });
 
+        const { protocol, hostname, port } = kbnTestConfig.getUrlPartsWithStrippedDefaultPort();
         expect(resp.hits.hits[0]._source?.ruleType).eql('observability.rules.custom_threshold');
         expect(resp.hits.hits[0]._source?.alertDetailsUrl).eql(
-          `https://localhost:5601/app/observability/alerts/${alertId}`
+          `${protocol}://${hostname}${port ? `:${port}` : ''}/app/observability/alerts/${alertId}`
         );
-
         expect(resp.hits.hits[0]._source?.reason).eql(
-          `Document count is 3, not between the threshold of 1 and 2. (duration: 1 min, data view: ${DATA_VIEW_NAME})`
+          'Average system.cpu.user.pct reported no data in the last 5m'
         );
-        expect(resp.hits.hits[0]._source?.value).eql('3');
+        expect(resp.hits.hits[0]._source?.value).eql('[NO DATA]');
 
         const parsedViewInAppUrl = parseSearchParams<LogsExplorerLocatorParsedParams>(
           new URL(resp.hits.hits[0]._source?.viewInAppUrl || '').search
@@ -263,7 +236,7 @@ export default function ({ getService }: FtrProviderContext) {
         expect(omit(parsedViewInAppUrl.params, 'timeRange.from')).eql({
           dataset: DATA_VIEW_ID,
           timeRange: { to: 'now' },
-          query: { query: 'host.name:* and container.id:*', language: 'kuery' },
+          query: { query: '', language: 'kuery' },
           filters: [],
         });
         expect(parsedViewInAppUrl.params.timeRange.from).match(ISO_DATE_REGEX);
