+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/alerting/custom_threshold/avg_us_fired.ts
@@ -5,97 +5,86 @@
  * 2.0.
  */
 
-import { cleanup, generate, Dataset, PartialConfig } from '@kbn/data-forge';
+import moment from 'moment';
+import { format } from 'url';
+import expect from '@kbn/expect';
+import { COMPARATORS } from '@kbn/alerting-comparators';
+import { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
 import { Aggregators } from '@kbn/observability-plugin/common/custom_threshold_rule/types';
 import { FIRED_ACTIONS_ID } from '@kbn/observability-plugin/server/lib/rules/custom_threshold/constants';
-import expect from '@kbn/expect';
 import { OBSERVABILITY_THRESHOLD_RULE_TYPE_ID } from '@kbn/rule-data-utils';
-import { parseSearchParams } from '@kbn/share-plugin/common/url_service';
-import { omit } from 'lodash';
-import { COMPARATORS } from '@kbn/alerting-comparators';
 import { kbnTestConfig } from '@kbn/test';
-import { FtrProviderContext } from '../../../ftr_provider_context';
-import { ISO_DATE_REGEX } from './constants';
-import { ActionDocument, LogsExplorerLocatorParsedParams } from './typings';
-import type { InternalRequestHeader, RoleCredentials } from '../../../../shared/services';
+import type { InternalRequestHeader, RoleCredentials } from '@kbn/ftr-common-functional-services';
+import { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
+import { getSyntraceClient, generateData } from './helpers/syntrace';
+import { ActionDocument } from './types';
 
-export default function ({ getService }: FtrProviderContext) {
+export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
+  const start = moment(Date.now()).subtract(10, 'minutes').valueOf();
+  const end = moment(Date.now()).add(15, 'minutes').valueOf();
   const esClient = getService('es');
-  const supertest = getService('supertest');
+  const samlAuth = getService('samlAuth');
+  const supertestWithoutAuth = getService('supertestWithoutAuth');
   const esDeleteAllIndices = getService('esDeleteAllIndices');
   const alertingApi = getService('alertingApi');
-  const logger = getService('log');
-  const svlUserManager = getService('svlUserManager');
-  const svlCommonApi = getService('svlCommonApi');
+  const dataViewApi = getService('dataViewApi');
+  const config = getService('config');
+  const kibanaServerConfig = config.get('servers.kibana');
+  const isServerless = config.get('serverless');
+  const expectedConsumer = isServerless ? 'observability' : 'logs';
+  const kibanaUrl = format(kibanaServerConfig);
   let roleAuthc: RoleCredentials;
   let internalReqHeader: InternalRequestHeader;
 
-  describe('Custom Threshold rule - P99 - PCT - FIRED', () => {
+  describe('AVG - US - FIRED', () => {
     const CUSTOM_THRESHOLD_RULE_ALERT_INDEX = '.alerts-observability.threshold.alerts-default';
     const ALERT_ACTION_INDEX = 'alert-action-threshold';
-    const DATA_VIEW_TITLE = 'kbn-data-forge-fake_hosts.fake_hosts-*';
-    const DATA_VIEW_NAME = 'ad-hoc-data-view-name';
+    const DATA_VIEW = 'traces-apm*,metrics-apm*,logs-apm*';
     const DATA_VIEW_ID = 'data-view-id';
-    const MOCKED_AD_HOC_DATA_VIEW = {
-      id: DATA_VIEW_ID,
-      title: DATA_VIEW_TITLE,
-      timeFieldName: '@timestamp',
-      sourceFilters: [],
-      fieldFormats: {},
-      runtimeFieldMap: {},
-      allowNoIndex: false,
-      name: DATA_VIEW_NAME,
-      allowHidden: false,
-    };
-    let dataForgeConfig: PartialConfig;
-    let dataForgeIndices: string[];
+    const DATA_VIEW_NAME = 'test-data-view-name';
+
+    let synthtraceEsClient: ApmSynthtraceEsClient;
     let actionId: string;
     let ruleId: string;
     let alertId: string;
 
     before(async () => {
-      roleAuthc = await svlUserManager.createM2mApiKeyWithRoleScope('admin');
-      internalReqHeader = svlCommonApi.getInternalRequestHeader();
-      dataForgeConfig = {
-        schedule: [
-          {
-            template: 'good',
-            start: 'now-10m',
-            end: 'now+5m',
-            metrics: [{ name: 'system.cpu.user.pct', method: 'linear', start: 2.5, end: 2.5 }],
-          },
-        ],
-        indexing: {
-          dataset: 'fake_hosts' as Dataset,
-          eventsPerCycle: 1,
-          interval: 10000,
-          alignEventsToInterval: true,
-        },
-      };
-      dataForgeIndices = await generate({ client: esClient, config: dataForgeConfig, logger });
-      logger.info(JSON.stringify(dataForgeIndices.join(',')));
-      await alertingApi.waitForDocumentInIndex({
-        indexName: DATA_VIEW_TITLE,
-        docCountTarget: 270,
+      roleAuthc = await samlAuth.createM2mApiKeyWithRoleScope('admin');
+      internalReqHeader = samlAuth.getInternalRequestHeader();
+      synthtraceEsClient = await getSyntraceClient({ esClient, kibanaUrl });
+      await generateData({ synthtraceEsClient, start, end });
+      await dataViewApi.create({
+        name: DATA_VIEW_NAME,
+        id: DATA_VIEW_ID,
+        title: DATA_VIEW,
+        roleAuthc,
       });
     });
 
     after(async () => {
-      await supertest.delete(`/api/alerting/rule/${ruleId}`).set(internalReqHeader);
-      await supertest.delete(`/api/actions/connector/${actionId}`).set(internalReqHeader);
+      await supertestWithoutAuth
+        .delete(`/api/alerting/rule/${ruleId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(internalReqHeader);
+      await supertestWithoutAuth
+        .delete(`/api/actions/connector/${actionId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(internalReqHeader);
+      await esDeleteAllIndices([ALERT_ACTION_INDEX]);
       await esClient.deleteByQuery({
         index: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
         query: { term: { 'kibana.alert.rule.uuid': ruleId } },
-        conflicts: 'proceed',
       });
       await esClient.deleteByQuery({
         index: '.kibana-event-log-*',
-        query: { term: { 'kibana.alert.rule.consumer': 'logs' } },
-        conflicts: 'proceed',
+        query: { term: { 'kibana.alert.rule.consumer': expectedConsumer } },
       });
-      await esDeleteAllIndices([ALERT_ACTION_INDEX, ...dataForgeIndices]);
-      await cleanup({ client: esClient, config: dataForgeConfig, logger });
-      await svlUserManager.invalidateM2mApiKeyWithRoleScope(roleAuthc);
+      await synthtraceEsClient.clean();
+      await dataViewApi.delete({
+        id: DATA_VIEW_ID,
+        roleAuthc,
+      });
+      await samlAuth.invalidateM2mApiKeyWithRoleScope(roleAuthc);
     });
 
     describe('Rule creation', () => {
@@ -109,17 +98,19 @@ export default function ({ getService }: FtrProviderContext) {
         const createdRule = await alertingApi.createRule({
           roleAuthc,
           tags: ['observability'],
-          consumer: 'observability',
+          consumer: expectedConsumer,
           name: 'Threshold rule',
           ruleTypeId: OBSERVABILITY_THRESHOLD_RULE_TYPE_ID,
           params: {
             criteria: [
               {
                 comparator: COMPARATORS.GREATER_THAN,
-                threshold: [0.5],
+                threshold: [7500000],
                 timeSize: 5,
                 timeUnit: 'm',
-                metrics: [{ name: 'A', field: 'system.cpu.user.pct', aggType: Aggregators.P99 }],
+                metrics: [
+                  { name: 'A', field: 'span.self_time.sum.us', aggType: Aggregators.AVERAGE },
+                ],
               },
             ],
             alertOnNoData: true,
@@ -129,7 +120,7 @@ export default function ({ getService }: FtrProviderContext) {
                 query: '',
                 language: 'kuery',
               },
-              index: MOCKED_AD_HOC_DATA_VIEW,
+              index: DATA_VIEW_ID,
             },
           },
           actions: [
@@ -143,8 +134,6 @@ export default function ({ getService }: FtrProviderContext) {
                     alertDetailsUrl: '{{context.alertDetailsUrl}}',
                     reason: '{{context.reason}}',
                     value: '{{context.value}}',
-                    host: '{{context.host}}',
-                    viewInAppUrl: '{{context.viewInAppUrl}}',
                   },
                 ],
               },
@@ -169,12 +158,6 @@ export default function ({ getService }: FtrProviderContext) {
         expect(executionStatus).to.be('active');
       });
 
-      it('should find the created rule with correct information about the consumer', async () => {
-        const match = await alertingApi.findInRules(roleAuthc, ruleId);
-        expect(match).not.to.be(undefined);
-        expect(match.consumer).to.be('observability');
-      });
-
       it('should set correct information in the alert document', async () => {
         const resp = await alertingApi.waitForAlertInIndex({
           indexName: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
@@ -186,7 +169,7 @@ export default function ({ getService }: FtrProviderContext) {
           'kibana.alert.rule.category',
           'Custom threshold'
         );
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', 'observability');
+        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', expectedConsumer);
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.name', 'Threshold rule');
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.producer', 'observability');
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.revision', 0);
@@ -208,56 +191,41 @@ export default function ({ getService }: FtrProviderContext) {
         expect(resp.hits.hits[0]._source).property('kibana.alert.workflow_status', 'open');
         expect(resp.hits.hits[0]._source).property('event.kind', 'signal');
         expect(resp.hits.hits[0]._source).property('event.action', 'open');
-
+        expect(resp.hits.hits[0]._source)
+          .property('kibana.alert.evaluation.threshold')
+          .eql([7500000]);
         expect(resp.hits.hits[0]._source)
           .property('kibana.alert.rule.parameters')
           .eql({
             criteria: [
               {
                 comparator: '>',
-                threshold: [0.5],
+                threshold: [7500000],
                 timeSize: 5,
                 timeUnit: 'm',
-                metrics: [{ name: 'A', field: 'system.cpu.user.pct', aggType: 'p99' }],
+                metrics: [{ name: 'A', field: 'span.self_time.sum.us', aggType: 'avg' }],
               },
             ],
             alertOnNoData: true,
             alertOnGroupDisappear: true,
-            searchConfiguration: {
-              index: MOCKED_AD_HOC_DATA_VIEW,
-              query: { query: '', language: 'kuery' },
-            },
+            searchConfiguration: { index: 'data-view-id', query: { query: '', language: 'kuery' } },
           });
       });
 
-      it('should set correct action variables', async () => {
+      it('should set correct action parameter: ruleType', async () => {
         const resp = await alertingApi.waitForDocumentInIndex<ActionDocument>({
           indexName: ALERT_ACTION_INDEX,
           docCountTarget: 1,
         });
-
-        const { protocol, hostname, port } = kbnTestConfig.getUrlParts();
+        const { protocol, hostname, port } = kbnTestConfig.getUrlPartsWithStrippedDefaultPort();
         expect(resp.hits.hits[0]._source?.ruleType).eql('observability.rules.custom_threshold');
         expect(resp.hits.hits[0]._source?.alertDetailsUrl).eql(
           `${protocol}://${hostname}${port ? `:${port}` : ''}/app/observability/alerts/${alertId}`
         );
         expect(resp.hits.hits[0]._source?.reason).eql(
-          `99th percentile of system.cpu.user.pct is 250%, above the threshold of 50%. (duration: 5 mins, data view: ${DATA_VIEW_NAME})`
+          `Average span.self_time.sum.us is 10,000,000, above the threshold of 7,500,000. (duration: 5 mins, data view: ${DATA_VIEW_NAME})`
         );
-        expect(resp.hits.hits[0]._source?.value).eql('250%');
-
-        const parsedViewInAppUrl = parseSearchParams<LogsExplorerLocatorParsedParams>(
-          new URL(resp.hits.hits[0]._source?.viewInAppUrl || '').search
-        );
-
-        expect(resp.hits.hits[0]._source?.viewInAppUrl).contain('LOGS_EXPLORER_LOCATOR');
-        expect(omit(parsedViewInAppUrl.params, 'timeRange.from')).eql({
-          dataset: DATA_VIEW_TITLE,
-          timeRange: { to: 'now' },
-          query: { query: '', language: 'kuery' },
-          filters: [],
-        });
-        expect(parsedViewInAppUrl.params.timeRange.from).match(ISO_DATE_REGEX);
+        expect(resp.hits.hits[0]._source?.value).eql('10,000,000');
       });
     });
   });
