+++ x-pack/test/api_integration/deployment_agnostic/apis/observability/alerting/custom_threshold/custom_eq_avg_bytes_fired.ts
@@ -4,111 +4,136 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
 
-import moment from 'moment';
-import { ApmSynthtraceEsClient } from '@kbn/apm-synthtrace';
-import { format } from 'url';
-import { Aggregators } from '@kbn/observability-plugin/common/custom_threshold_rule/types';
-import { FIRED_ACTIONS_ID } from '@kbn/observability-plugin/server/lib/rules/custom_threshold/constants';
 import expect from '@kbn/expect';
-import { OBSERVABILITY_THRESHOLD_RULE_TYPE_ID } from '@kbn/rule-data-utils';
+import { cleanup, Dataset, generate, PartialConfig } from '@kbn/data-forge';
+import { Aggregators } from '@kbn/observability-plugin/common/custom_threshold_rule/types';
 import { COMPARATORS } from '@kbn/alerting-comparators';
-import { FtrProviderContext } from '../../common/ftr_provider_context';
-import { createIndexConnector, createRule } from '../helpers/alerting_api_helper';
-import { createDataView, deleteDataView } from '../helpers/data_view';
-import { getSyntraceClient, generateData } from '../helpers/syntrace';
-import {
-  waitForAlertInIndex,
-  waitForDocumentInIndex,
-  waitForRuleStatus,
-} from '../helpers/alerting_wait_for_helpers';
-import { ActionDocument } from './typings';
+import { FIRED_ACTIONS_ID } from '@kbn/observability-plugin/server/lib/rules/custom_threshold/constants';
+import { OBSERVABILITY_THRESHOLD_RULE_TYPE_ID } from '@kbn/rule-data-utils';
+import { kbnTestConfig } from '@kbn/test';
+import type { InternalRequestHeader, RoleCredentials } from '@kbn/ftr-common-functional-services';
+import { DeploymentAgnosticFtrProviderContext } from '../../../../ftr_provider_context';
+import { ActionDocument } from './types';
 
-// eslint-disable-next-line import/no-default-export
-export default function ({ getService }: FtrProviderContext) {
-  const start = moment(Date.now()).subtract(10, 'minutes').valueOf();
-  const end = moment(Date.now()).valueOf();
+export default function ({ getService }: DeploymentAgnosticFtrProviderContext) {
   const esClient = getService('es');
+  const supertestWithoutAuth = getService('supertestWithoutAuth');
   const esDeleteAllIndices = getService('esDeleteAllIndices');
-  const config = getService('config');
-  const kibanaServerConfig = config.get('servers.kibana');
-  const kibanaUrl = format(kibanaServerConfig);
-  const supertest = getService('supertest');
   const logger = getService('log');
-  const retryService = getService('retry');
+  const alertingApi = getService('alertingApi');
+  const dataViewApi = getService('dataViewApi');
+  const samlAuth = getService('samlAuth');
+  let roleAuthc: RoleCredentials;
+  let internalReqHeader: InternalRequestHeader;
+  const config = getService('config');
+  const isServerless = config.get('serverless');
+  const expectedConsumer = isServerless ? 'observability' : 'logs';
 
-  describe('Custom Threshold rule - AVG - US - FIRED', () => {
+  describe('CUSTOM_EQ - AVG - BYTES - FIRED', () => {
     const CUSTOM_THRESHOLD_RULE_ALERT_INDEX = '.alerts-observability.threshold.alerts-default';
+    const DATA_VIEW = 'kbn-data-forge-fake_hosts.fake_hosts-*';
     const ALERT_ACTION_INDEX = 'alert-action-threshold';
-    const DATA_VIEW = 'traces-apm*,metrics-apm*,logs-apm*';
     const DATA_VIEW_ID = 'data-view-id';
-    const DATA_VIEW_NAME = 'test-data-view-name';
-
-    let synthtraceEsClient: ApmSynthtraceEsClient;
+    let dataForgeConfig: PartialConfig;
+    let dataForgeIndices: string[];
     let actionId: string;
     let ruleId: string;
     let alertId: string;
 
     before(async () => {
-      synthtraceEsClient = await getSyntraceClient({ esClient, kibanaUrl });
-      await generateData({ synthtraceEsClient, start, end });
-      await createDataView({
-        supertest,
-        name: DATA_VIEW_NAME,
+      roleAuthc = await samlAuth.createM2mApiKeyWithRoleScope('admin');
+      internalReqHeader = samlAuth.getInternalRequestHeader();
+      dataForgeConfig = {
+        schedule: [
+          {
+            template: 'good',
+            start: 'now-15m',
+            end: 'now+5m',
+            metrics: [
+              { name: 'system.cpu.user.pct', method: 'linear', start: 2.5, end: 2.5 },
+              { name: 'system.cpu.total.pct', method: 'linear', start: 0.5, end: 0.5 },
+              { name: 'system.cpu.total.norm.pct', method: 'linear', start: 0.8, end: 0.8 },
+            ],
+          },
+        ],
+        indexing: {
+          dataset: 'fake_hosts' as Dataset,
+          eventsPerCycle: 1,
+          interval: 10000,
+          alignEventsToInterval: true,
+        },
+      };
+      dataForgeIndices = await generate({ client: esClient, config: dataForgeConfig, logger });
+      await alertingApi.waitForDocumentInIndex({ indexName: DATA_VIEW, docCountTarget: 360 });
+      await dataViewApi.create({
+        name: DATA_VIEW,
         id: DATA_VIEW_ID,
         title: DATA_VIEW,
-        logger,
+        roleAuthc,
       });
     });
 
     after(async () => {
-      await supertest.delete(`/api/alerting/rule/${ruleId}`).set('kbn-xsrf', 'foo');
-      await supertest.delete(`/api/actions/connector/${actionId}`).set('kbn-xsrf', 'foo');
-      await esDeleteAllIndices([ALERT_ACTION_INDEX]);
+      await supertestWithoutAuth
+        .delete(`/api/alerting/rule/${ruleId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(internalReqHeader);
+      await supertestWithoutAuth
+        .delete(`/api/actions/connector/${actionId}`)
+        .set(roleAuthc.apiKeyHeader)
+        .set(internalReqHeader);
       await esClient.deleteByQuery({
         index: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
         query: { term: { 'kibana.alert.rule.uuid': ruleId } },
+        conflicts: 'proceed',
       });
       await esClient.deleteByQuery({
         index: '.kibana-event-log-*',
-        query: { term: { 'kibana.alert.rule.consumer': 'logs' } },
+        query: { term: { 'rule.id': ruleId } },
+        conflicts: 'proceed',
       });
-      await synthtraceEsClient.clean();
-      await deleteDataView({
-        supertest,
+      await dataViewApi.delete({
         id: DATA_VIEW_ID,
-        logger,
+        roleAuthc,
       });
+      await esDeleteAllIndices([ALERT_ACTION_INDEX, ...dataForgeIndices]);
+      await cleanup({ client: esClient, config: dataForgeConfig, logger });
+      await samlAuth.invalidateM2mApiKeyWithRoleScope(roleAuthc);
     });
 
     describe('Rule creation', () => {
       it('creates rule successfully', async () => {
-        actionId = await createIndexConnector({
-          supertest,
+        actionId = await alertingApi.createIndexConnector({
+          roleAuthc,
           name: 'Index Connector: Threshold API test',
           indexName: ALERT_ACTION_INDEX,
-          logger,
         });
 
-        const createdRule = await createRule({
-          supertest,
-          logger,
-          esClient,
+        const createdRule = await alertingApi.createRule({
+          roleAuthc,
           tags: ['observability'],
-          consumer: 'logs',
+          consumer: expectedConsumer,
           name: 'Threshold rule',
           ruleTypeId: OBSERVABILITY_THRESHOLD_RULE_TYPE_ID,
           params: {
             criteria: [
               {
-                aggType: 'custom',
                 comparator: COMPARATORS.GREATER_THAN,
-                threshold: [7500000],
-                timeSize: 5,
+                threshold: [0.9],
+                timeSize: 1,
                 timeUnit: 'm',
                 metrics: [
-                  { name: 'A', field: 'span.self_time.sum.us', aggType: Aggregators.AVERAGE },
+                  { name: 'A', field: 'system.network.in.bytes', aggType: Aggregators.AVERAGE },
+                  { name: 'B', field: 'system.network.out.bytes', aggType: Aggregators.AVERAGE },
                 ],
+                equation: '((A + A) / (B + B))',
               },
             ],
             alertOnNoData: true,
@@ -148,23 +173,24 @@ export default function ({ getService }: FtrProviderContext) {
       });
 
       it('should be active', async () => {
-        const executionStatus = await waitForRuleStatus({
-          id: ruleId,
+        const executionStatus = await alertingApi.waitForRuleStatus({
+          roleAuthc,
+          ruleId,
           expectedStatus: 'active',
-          supertest,
-          retryService,
-          logger,
         });
-        expect(executionStatus.status).to.be('active');
+        expect(executionStatus).to.be('active');
+      });
+
+      it('should find the created rule with correct information about the consumer', async () => {
+        const match = await alertingApi.findInRules(roleAuthc, ruleId);
+        expect(match).not.to.be(undefined);
+        expect(match.consumer).to.be(expectedConsumer);
       });
 
       it('should set correct information in the alert document', async () => {
-        const resp = await waitForAlertInIndex({
-          esClient,
+        const resp = await alertingApi.waitForAlertInIndex({
           indexName: CUSTOM_THRESHOLD_RULE_ALERT_INDEX,
           ruleId,
-          retryService,
-          logger,
         });
         alertId = (resp.hits.hits[0]._source as any)['kibana.alert.uuid'];
 
@@ -172,7 +198,7 @@ export default function ({ getService }: FtrProviderContext) {
           'kibana.alert.rule.category',
           'Custom threshold'
         );
-        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', 'logs');
+        expect(resp.hits.hits[0]._source).property('kibana.alert.rule.consumer', expectedConsumer);
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.name', 'Threshold rule');
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.producer', 'observability');
         expect(resp.hits.hits[0]._source).property('kibana.alert.rule.revision', 0);
@@ -194,20 +220,21 @@ export default function ({ getService }: FtrProviderContext) {
         expect(resp.hits.hits[0]._source).property('kibana.alert.workflow_status', 'open');
         expect(resp.hits.hits[0]._source).property('event.kind', 'signal');
         expect(resp.hits.hits[0]._source).property('event.action', 'open');
-        expect(resp.hits.hits[0]._source)
-          .property('kibana.alert.evaluation.threshold')
-          .eql([7500000]);
+        expect(resp.hits.hits[0]._source).property('kibana.alert.evaluation.threshold').eql([0.9]);
         expect(resp.hits.hits[0]._source)
           .property('kibana.alert.rule.parameters')
           .eql({
             criteria: [
               {
-                aggType: 'custom',
-                comparator: '>',
-                threshold: [7500000],
-                timeSize: 5,
+                comparator: COMPARATORS.GREATER_THAN,
+                threshold: [0.9],
+                timeSize: 1,
                 timeUnit: 'm',
-                metrics: [{ name: 'A', field: 'span.self_time.sum.us', aggType: 'avg' }],
+                metrics: [
+                  { name: 'A', field: 'system.network.in.bytes', aggType: Aggregators.AVERAGE },
+                  { name: 'B', field: 'system.network.out.bytes', aggType: Aggregators.AVERAGE },
+                ],
+                equation: '((A + A) / (B + B))',
               },
             ],
             alertOnNoData: true,
@@ -216,22 +243,21 @@ export default function ({ getService }: FtrProviderContext) {
           });
       });
 
-      it('should set correct action parameter: ruleType', async () => {
-        const resp = await waitForDocumentInIndex<ActionDocument>({
-          esClient,
+      it('should set correct action variables', async () => {
+        const resp = await alertingApi.waitForDocumentInIndex<ActionDocument>({
           indexName: ALERT_ACTION_INDEX,
-          retryService,
-          logger,
+          docCountTarget: 1,
         });
 
+        const { protocol, hostname, port } = kbnTestConfig.getUrlParts();
         expect(resp.hits.hits[0]._source?.ruleType).eql('observability.rules.custom_threshold');
         expect(resp.hits.hits[0]._source?.alertDetailsUrl).eql(
-          `https://localhost:5601/app/observability/alerts/${alertId}`
+          `${protocol}://${hostname}${port ? `:${port}` : ''}/app/observability/alerts/${alertId}`
         );
         expect(resp.hits.hits[0]._source?.reason).eql(
-          `Average span.self_time.sum.us is 10,000,000, above the threshold of 7,500,000. (duration: 5 mins, data view: ${DATA_VIEW_NAME})`
+          `Custom equation is 1 B, above the threshold of 0.9 B. (duration: 1 min, data view: ${DATA_VIEW})`
         );
-        expect(resp.hits.hits[0]._source?.value).eql('10,000,000');
+        expect(resp.hits.hits[0]._source?.value).eql('1 B');
       });
     });
   });
