+++ x-pack/test_serverless/api_integration/test_suites/observability/index.feature_flags.ts
@@ -9,8 +9,6 @@ import { FtrProviderContext } from '../../ftr_provider_context';
 
 export default function ({ loadTestFile }: FtrProviderContext) {
   describe('Serverless observability API - feature flags', function () {
-    loadTestFile(require.resolve('./custom_threshold_rule'));
-    loadTestFile(require.resolve('./infra'));
     loadTestFile(require.resolve('./platform_security'));
     loadTestFile(require.resolve('../common/platform_security/roles_routes_feature_flag.ts'));
     loadTestFile(require.resolve('../common/management/multiple_spaces_enabled.ts'));
