+++ x-pack/test_serverless/tsconfig.json
@@ -32,9 +32,6 @@
     "@kbn/telemetry-plugin",
     "@kbn/telemetry-collection-xpack-plugin",
     "@kbn/telemetry-tools",
-    "@kbn/infra-plugin",
-    "@kbn/observability-plugin",
-    "@kbn/data-forge",
     "@kbn/ftr-common-functional-services",
     "@kbn/apm-plugin",
     "@kbn/server-route-repository",
@@ -83,12 +80,9 @@
     "@kbn/reporting-common",
     "@kbn/es-query",
     "@kbn/slo-plugin",
-    "@kbn/share-plugin",
     "@kbn/es-query",
-    "@kbn/utility-types",
     "@kbn/synthetics-plugin",
     "@kbn/dataset-quality-plugin",
-    "@kbn/alerting-comparators",
     "@kbn/search-types",
     "@kbn/config-schema",
     "@kbn/features-plugin",
